requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-otp==1.2.4
django-ratelimit==4.1.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==37.5.3
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.4.3
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.26.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.6.0
drf-yasg
pillow
